requirements.txt
Flask>=2.0
Pillow>=9.0
torch>=2.1.0
sentencepiece>=0.1.99
transformers>=4.30.0
numpy>=1.23.0
pandas>=1.5.0
firebase-admin>=6.0.0
scikit-learn>=1.2.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
gunicorn>=21.2.0